+++ popup/scripts/subject.js
@@ -8,8 +8,35 @@ if (!messagePopup) {
 const url = new URL(window.location.href);
 let id = url.searchParams.get("id");
 
-chrome.storage.local.get(["settings"], async result => {
+let fetchKeys = ["settings"];
+if (id == "rand")
+	fetchKeys = [...fetchKeys, "kana_vocab_assoc", "kanji_assoc", "vocabulary_assoc"];
+else if (id == "rand-kanji")
+	fetchKeys = [...fetchKeys, "kanji_assoc"];
+else if (id == "rand-vocabulary")
+	fetchKeys = [...fetchKeys, "vocabulary_assoc", "kana_vocab_assoc"];
+
+chrome.storage.local.get(fetchKeys, async result => {
 	if (popupLoading) popupLoading.remove();
+
+	if (id.includes("rand")) {
+		let assocs = {};
+		if (id == "rand") {
+			assocs = {...result["kana_vocab_assoc"], ...result["kanji_assoc"], ...result["vocabulary_assoc"]};
+		}
+		else if (id == "rand-kanji") {
+			assocs = result["kanji_assoc"];
+		}
+		else if (id == "rand-vocabulary") {
+			assocs = {...result["vocabulary_assoc"], ...result["kana_vocab_assoc"]};
+		}
+		
+		const ids = Object.values(assocs);
+		// get random id
+		id = ids[Math.floor(Math.random() * ids.length)];
+		// update url
+		history.replaceState(null, null, `?id=${id}`);
+	} 
 	
 	const highlightStyleSettings = result["settings"]["highlight_style"];
 	let highlightingClass, notLearnedHighlightingClass;
@@ -40,6 +67,7 @@ chrome.storage.local.get(["settings"], async result => {
 			autoplayAudio: result["settings"]["kanji_details_popup"]["audio_autoplay"],
 		}
 	);
+	console.log("details", details);
 
 	await details.create(true);